complete new goal register

+++ src/app/goal/goal.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, Output, EventEmitter } from '@angular/core';
 import { Goal } from '../goal';
-import { RouterModule, Router} from '@angular/router';
+import { RouterModule, Router } from '@angular/router';
 import { GoalService } from '../goal.service';
 
 
@@ -11,20 +11,22 @@ import { GoalService } from '../goal.service';
   styleUrl: './goal.component.css'
 })
 export class GoalComponent {
-  @Input() goal!:Goal;  //non-null assertion operator
+  @Input() goal!: Goal;  //non-null assertion operator
   @Output() goalDeleted = new EventEmitter<string>();
-  constructor(private router:Router, private goalService:GoalService){
+  constructor(private router: Router, private goalService: GoalService) {
 
   }
 
-  seeDetails(){
+  seeDetails() {
     console.log('Calling by method!')
     this.router.navigate(['details', this.goal.id])
   }
 
-  deleteGoal(){
-    this.goalService.delete(this.goal.id).then( ()=>{
-       this.goalDeleted.emit (this.goal.id)
-    })
+  deleteGoal() {
+    if (this.goal.id) {
+      this.goalService.delete(this.goal.id).then(() => {
+        this.goalDeleted.emit(this.goal.id)
+      })
+    }
   }
 }
